feat: refact backend status
